Added configuration !

+++ src/main/java/me/nik/bootcleaner/Main.java
@@ -1,11 +1,13 @@
 package me.nik.bootcleaner;
 
+import me.nik.bootcleaner.config.ConfigurationBuilder;
 import me.nik.bootcleaner.enums.Directories;
 import me.nik.bootcleaner.utils.FileUtils;
 import me.nik.bootcleaner.utils.MiscUtils;
 
 import javax.swing.*;
 import java.io.File;
+import java.io.IOException;
 import java.util.Arrays;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
@@ -14,11 +16,25 @@ public class Main {
 
     public static void main(String[] args) {
 
-        MiscUtils.executeCmdCommand("cmd /C start ipconfig /flushdns");
-        MiscUtils.executeCmdCommand("cmd /C rd /s /q %systemdrive%\\$RECYCLE.BIN");
+        //Load the configuration file
+        try {
+            new ConfigurationBuilder(Config.class, new File("config.yml")).build(true);
+        } catch (IOException e) {
+            JOptionPane.showMessageDialog(null,
+                    "Configuration Error, " + e.getLocalizedMessage(),
+                    "BootCleaner",
+                    JOptionPane.ERROR_MESSAGE);
+
+            System.exit(1);
+        }
+
+        if (Config.cleanDnsCache) MiscUtils.executeCmdCommand("cmd /C start ipconfig /flushdns");
+
+        if (Config.emptyRecycleBin) MiscUtils.executeCmdCommand("cmd /C rd /s /q %systemdrive%\\$RECYCLE.BIN");
 
         int cleanedFiles;
 
+        //Execute the cleaning asynchronously
         try {
 
             cleanedFiles = CompletableFuture
@@ -30,15 +46,18 @@ public class Main {
 
             JOptionPane.showMessageDialog(null,
                     "Couldn't clean temporary files asynchronously",
-                    "BootCleaner by Nik",
+                    "BootCleaner",
                     JOptionPane.ERROR_MESSAGE);
 
             return;
         }
 
-        JOptionPane.showMessageDialog(null,
-                "Cleaned " + cleanedFiles + " temporary files",
-                "BootCleaner by Nik",
-                JOptionPane.PLAIN_MESSAGE);
+        if (Config.showMessage) {
+
+            JOptionPane.showMessageDialog(null,
+                    "Cleaned " + cleanedFiles + " temporary files",
+                    "BootCleaner",
+                    JOptionPane.PLAIN_MESSAGE);
+        }
     }
 }
